+++ packages/@expo/config/src/paths/paths.ts
@@ -4,7 +4,6 @@ import resolveFrom from 'resolve-from';
 
 import { getConfig } from '../Config';
 import { ProjectConfig } from '../Config.types';
-import { ConfigError } from '../Errors';
 import { getBareExtensions } from './extensions';
 
 // https://github.com/facebook/create-react-app/blob/9750738cce89a967cc71f28390daf5d4311b193c/packages/react-scripts/config/paths.js#L22
@@ -61,15 +60,7 @@ export function getEntryPointWithExtensions(
     }
   }
 
-  const { exp, pkg } = projectConfig;
-
-  if (typeof exp?.entryPoint === 'string') {
-    // We want to stop reading the app.json for determining the entry file in SDK +49
-    throw new ConfigError(
-      'expo.entryPoint has been removed in favor of the main field in the package.json.',
-      'DEPRECATED'
-    );
-  }
+  const { pkg } = projectConfig;
 
   if (pkg) {
     // If the config doesn't define a custom entry then we want to look at the `package.json`s `main` field, and try again.