+++ packages/@expo/config/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### 🐛 Bug fixes
 
+- Drop `entryPoint` usage.
 - Make `exp` type optional in `resolveEntryPoint`. ([#21643](https://github.com/expo/expo/pull/21643) by [@EvanBacon](https://github.com/EvanBacon))
 
 ### 💡 Others